+++ examples/data-sources/meraki_appliance_inbound_cellular_firewall_rules/data-source.tf
@@ -1,0 +1,3 @@
+data "meraki_appliance_inbound_cellular_firewall_rules" "example" {
+  network_id = "L_123456"
+}